Changed home page route and links related
