+++ apps/web-app/pages/index.tsx
@@ -1,5 +1,41 @@
-import HomeTemplate from "../templates/Home"
+import { GetServerSideProps } from "next"
+import axios from "axios"
 
-const Home = () => <HomeTemplate />
+import { EventsDTO, SessionsDTO } from "../types"
+import CalendarPage from "../templates/CalendarPage"
 
-export default Home
+type Props = {
+    events: EventsDTO[]
+    sessions: SessionsDTO[]
+}
+
+export default function Event({ sessions, events }: Props) {
+    return <CalendarPage sessions={sessions} events={events} />
+}
+
+export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
+    try {
+        const url = process.env.URL_TO_FETCH
+
+        const eventsResponse = await fetch(`${url}/api/fetchEvents`)
+
+        const events: EventsDTO[] = await eventsResponse.json()
+
+        const responseSessions = await axios.get(`${url}/api/fetchSessions`, {
+            headers: {
+                Cookie: req.headers.cookie || "" // Pass cookies from the incoming request
+            }
+        })
+
+        const sessions = await responseSessions.data
+
+        return {
+            props: { sessions, events }
+        }
+    } catch (error) {
+        res.statusCode = 404
+        return {
+            props: {}
+        }
+    }
+}
